src/server-bt-django/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
supabase==2.3.4
PyJWT==2.8.0
APScheduler==3.10.4
requests==2.31.0
